+++ packages/react/src/components/color-picker/ColorPicker.tsx
@@ -1,70 +1,54 @@
 import * as React from 'react';
-import {ChromePicker, ChromePickerProps, ColorResult} from 'react-color';
-import {connect} from 'react-redux';
-import {isString, noop, uniqueId} from 'underscore';
-import {PlasmaState} from '../../PlasmaState';
+import {ChromePicker, ChromePickerProps, ColorChangeHandler, ColorResult} from 'react-color';
+import {useDispatch} from 'react-redux';
+import {uniqueId} from 'underscore';
 import {IDispatch} from '../../utils/ReduxUtils';
 import {changeInputValue} from '../input/InputActions';
 import {InputConnected} from '../input/Input';
-import {InputSelectors} from '../input/InputSelectors';
 
 export interface IColorPickerProps extends ChromePickerProps {
+    /**
+     * Id of the element
+     */
     id: string;
+    /**
+     * Hex value of the color
+     */
     defaultColor?: string;
 }
 
-const mapStateToProps = (state: PlasmaState, ownProps: IColorPickerProps) => {
-    const color = InputSelectors.getValue(state, {id: ownProps.id}) || ownProps.defaultColor;
-    return {
-        color,
+export const ColorPicker: React.FunctionComponent<IColorPickerProps> = ({
+    id = uniqueId('colorpicker'),
+    defaultColor = '#000',
+    onChangeComplete,
+    onChange,
+    ...props
+}) => {
+    const [color, setColor] = React.useState(defaultColor);
+    const dispatch: IDispatch = useDispatch();
+
+    React.useEffect(() => {
+        // sync the state on the first render
+        dispatch(changeInputValue(id, defaultColor, true));
+    }, []);
+
+    const onColorChange: ColorChangeHandler = (colorChanged, event) => {
+        setColor(colorChanged.hex);
+        onChange?.(colorChanged, event);
     };
+
+    const onUpdate: ColorChangeHandler = React.useCallback(
+        (colorPicked: ColorResult, event: React.ChangeEvent<HTMLInputElement>) => {
+            onChangeComplete?.(colorPicked, event);
+            dispatch(changeInputValue(id, colorPicked.hex, true));
+        },
+        []
+    );
+
+    return (
+        <>
+            <ChromePicker {...props} color={color} onChange={onColorChange} onChangeComplete={onUpdate} />
+            <InputConnected id={id} style={{display: 'none'}} defaultValue={defaultColor} />
+        </>
+    );
 };
-
-const mapDispatchToProps = (dispatch: IDispatch, ownProps: IColorPickerProps) => ({
-    onChangeComplete: (colorPicked: ColorResult, event: React.ChangeEvent<HTMLInputElement>) => {
-        ownProps.onChangeComplete?.(colorPicked, event);
-        dispatch(changeInputValue(ownProps.id, colorPicked.hex, true));
-    },
-});
-
-class ColorPickerDisconnected extends React.Component<
-    IColorPickerProps & ReturnType<typeof mapDispatchToProps> & ReturnType<typeof mapStateToProps>
-> {
-    static defaultProps = {
-        id: uniqueId('colorpicker'),
-        onChangeComplete: noop as any,
-    };
-
-    get colorForInput() {
-        if (this.props.color && isString(this.props.color)) {
-            return this.props.color;
-        }
-
-        if (this.props.defaultColor) {
-            return this.props.defaultColor;
-        }
-    }
-
-    render() {
-        return (
-            <>
-                <ChromePicker {...this.props} />
-                <InputConnected
-                    id={this.props.id}
-                    style={{display: 'none'}}
-                    defaultValue={this.colorForInput}
-                    value={this.colorForInput}
-                />
-            </>
-        );
-    }
-}
-
-export const ColorPicker = connect<
-    ReturnType<typeof mapStateToProps>,
-    ReturnType<typeof mapDispatchToProps>,
-    IColorPickerProps
->(
-    mapStateToProps,
-    mapDispatchToProps
-)(ColorPickerDisconnected as any);